Bug Fix: EFileWrite was not being saved
During modernisation renaming, an assignment to m_uiFileMode was
changed to a_fileMode instead of m_fileMode, causing file sending to
fail on RADRunner.  To make things worse, writes were not being
checked when saving the file, so a zero length file was written and
no error was displayed.

This is proof that you shouldn't touch working code!