Add RRemoteFactory::useLocal()
Added a method that allows an instance of RRemoteFactory that is
configured to access a remote RADRunner server to temporarily be used
for accessing local files.  This is useful, for instance, in Brunel
when a file tree is configured for remote access, but the user wishes
to perform a single operation on the local file system.