Fail build if Windows build or test fails
Until now, due to a misunderstanding of Windows PowerShell scripting,
if the Windows build or test failed, the GitHub workflow would not
fail.  This meant that some failing tests were not being picked up.
This has now been remedied, and the failing tests have also been fixed
so that they can pass the workflow.