Get the Qt version working on Windows again
With the upgrade to Qt6, the Windows version of Brunel stopped
exiting, and it's difficult to see how it ever worked to begin with,
even with Qt5. This commit also addresses a problem whereby Brunel
would ask twice whether to save a file that had been edited, due to
changes in the way the quit functionality is implemented on Qt6.