Add stack checking to MungWall
Whenever an allocation is freed by MungWall, the stack pointer will
be checked to ensure that it is within the stack bounds. Also, the
largest amount of stack that has been used will be tracked, so that
the developer can be notified if the stack is getting close to
overflowing.