Add support for catching signals
Until now, signal catching was performed in individual client programs.
However, because using signal() was found to not work well with the
so-called "slow" function calls such as select(), sigaction() is now
used, and the signal handling code was moved into the StdFuncs
library, so that all programs handle it consistently.