Don't pass entries array to RDir::read()
This was causing confusion in terms of ownership, and could lead to
surprises when RDir::close() was called and a list used by client code
would suddenly as a result become empty. Instead of this list pointer
being passed in, a pointer to the list of files can now be obtained by
calling RDir::getFiles().