StdGadgetTree::setContent() can now update existing lists
To suport dynamic addition or deletion of items to the list of items
displayed in the gadget tree, it is now possible to pass in a list ID
when calling setContent(). If no list ID is passed in, a new list will
be created as usual, but if a list ID is passed in, the list with that
ID will be updated with the new content, instead of a new one being
created.