Add support for watching individual files
Watching an entire directory for changes to a file is not guaranteed to
work on all operating systems. Some operating systems only notify that
a file is added or deleted, not changed. This change adds support for
watching an individual file (such as a configuration file) to be sure
that changes to it are picked up.