Add support for handling SIGBREAKF_CTRL_C
The Unix and Windows versions of this code already react to shutdown
messages sent from the system (i.e. Activity Monitor on Mac OS and the
Task Manager on Windows), but the Amiga OS version requires explicit
handling of this in the form of checking for the SIGBREAKF_CTRL_C
signal and handling it.